login & register mvc

+++ Session 19/controller/product_controller.php
@@ -19,18 +19,30 @@
 						if (!isset($_SESSION['username'])) {
 								header("Location: login.php");
 							}
+						if ($_SESSION['role'] != 'admin') {
+							echo "Bạn không có quyền này";
+							exit();
+						}
 						$this->deleteProduct();
 						break;
 					case 'productsEdit':
 						if (!isset($_SESSION['username'])) {
 								header("Location: login.php");
 							}
+						if ($_SESSION['role'] != 'admin') {
+							echo "Bạn không có quyền này";
+							exit();
+						}
 						$this->editProduct();
 						break;
 					case 'Add':
 						if (!isset($_SESSION['username'])) {
 								header("Location: login.php");
 							}
+						if ($_SESSION['role'] != 'admin') {
+							echo "Bạn không có quyền này";
+							exit();
+						}
 						$this->addProduct();
 						break;
 				}
